Corrección de Delete.
Solo borraba los valores especificados en la columna de condition, en el resto no

+++ MiniSQLServer/src/main/java/eus/ehu/giigsi/abd/parser/Delete.java
@@ -6,7 +6,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
 
-public class Delete implements MiniSQLQuery{
+public class Delete implements MiniSQLQuery {
     @Getter
     @Setter(AccessLevel.PRIVATE)
     public String table;
@@ -14,19 +14,22 @@ public class Delete implements MiniSQLQuery{
     @Setter(AccessLevel.PRIVATE)
     public Condition where;
 
-    public Delete(String table, Condition where)
-    {
+    public Delete(String table, Condition where) {
         this.table = table;
         this.where = where;
     }
 
-    public String execute(Database database)
-    {
-        boolean resultado = database.deleteWhere(table,where) ;
+    public String execute(Database database) {
+        if (table == null || where == null) {
+            return Constants.ERROR + "tabla o valores estan vacios";
 
-        if (resultado == true){
-            return Constants.DELETE_SUCCESS;
+        } else {
+            boolean resultado = database.deleteWhere(table, where);
+
+            if (resultado == true) {
+                return Constants.DELETE_SUCCESS;
+            }
+            return Constants.ERROR + "tabla o valores estan vacios";
         }
-        return Constants.ERROR;
     }
 }
